Decode SNMPv2 VarBind exception-values (GetRequest NoSuchObject, NoSuchInstance) in responses.
A response may have multiple VarBinds, some of which contain valid values, and others which contain exceptions.
We do not necessarily want to throw away the entire response as an error if some of the values are exceptions.

This changes the API for responses containing NoSuchObject or NoSuchInstance values, changing them to match the behavior of EndOfMibView.
Previously NoSuchObject decoded as UnsupportedBerType, and NoSuchInstance was a decoding error.